feat: implement comprehensive security hardening and compliance fixes
This commit introduces significant security and cryptographic robustness improvements to the pykyber library:

- Mitigated timing side-channel attacks by replacing value-dependent division operators with constant-time modular arithmetic in polynomial compression and message extraction functions.
- Integrated the zeroize crate to ensure that sensitive cryptographic material, including secret keys and intermediate state buffers, is explicitly wiped from memory when dropped.
- Achieved full NIST FIPS 203 compliance for Kyber-1024 by correcting the compression parameters (du=11, dv=5) and bit-packing logic, resulting in a standard-compliant 1568-byte ciphertext.
- Hardened the Kyber-512 implementation by updating the noise parameter eta1 to 3 and utilizing the correct CBD3 noise distribution as specified in the standard.
- Improved arithmetic robustness in the NTT implementation by using 32-bit intermediate types to prevent potential overflows and refactored Keccak squeezing to ensure memory safety for arbitrary output lengths.
- Added a full suite of verification tests and updated benchmarks to reflect the new security-first architecture.

+++ src/kyber/poly.rs
@@ -185,7 +185,7 @@ pub fn poly_tomont(r: &mut Poly) {
 
 pub fn poly_reduce(r: &mut Poly) {
     for i in 0..KYBER_N {
-        r.coeffs[i] = barrett_reduce(r.coeffs[i]);
+        r.coeffs[i] = barrett_reduce(r.coeffs[i] as i32);
     }
 }
